utils/requirements.txt
numpy
seaborn
nibabel
matplotlib
torchinfo
logging
skimage